Conflicts:
	src/include/Makefile.am
	src/include/pbs_nodes.h
	src/lib/Libutils/u_groups.c
	src/server/node_func.c
	src/test/node_func/Makefile.am
	src/test/node_manager/scaffolding.c
	src/test/node_manager/test_uut.c
	src/test/pbsd_init/scaffolding.c
	src/test/process_alps_status/scaffolding.c
	src/test/process_mom_update/scaffolding.c
	src/test/req_runjob/scaffolding.c
	src/test/start_exec/scaffolding.c
	src/test/svr_jobfunc/scaffolding.c

+++ src/test/attr_node_func/test_uut.c
@@ -21,16 +21,22 @@ START_TEST(test_one)
   int rc = decode_utc(&pa,"TTL",NULL,str,0);
   fail_unless(rc == 0);
   memset(&pa,0,sizeof(pa));
-  sprintf(str,"%04d-%02d-%02dT%02d:%02d:%02d%c%02d",tm.tm_year+1900,tm.tm_mon+1,tm.tm_mday,tm.tm_hour,tm.tm_min,tm.tm_sec,tm.tm_gmtoff < 0 ? '-':'+',abs(tm.tm_gmtoff)/3600);
+  sprintf(str,"%04d-%02d-%02dT%02d:%02d:%02d%c%02d",
+    tm.tm_year+1900, tm.tm_mon+1, tm.tm_mday, tm.tm_hour, tm.tm_min, tm.tm_sec,
+    tm.tm_gmtoff < 0 ? '-':'+', (int)abs(tm.tm_gmtoff)/3600);
   rc = decode_utc(&pa,"TTL",NULL,str,0);
   fail_unless(rc == 0);
   memset(&pa,0,sizeof(pa));
-  sprintf(str,"%04d-%02d-%02dT%02d:%02d:%02d%c%02d%02d",tm.tm_year+1900,tm.tm_mon+1,tm.tm_mday,tm.tm_hour,tm.tm_min,tm.tm_sec,tm.tm_gmtoff < 0 ? '-':'+',abs(tm.tm_gmtoff)/3600,(abs(tm.tm_gmtoff)/60)%60);
+  sprintf(str,"%04d-%02d-%02dT%02d:%02d:%02d%c%02d%02d",
+    tm.tm_year+1900, tm.tm_mon+1, tm.tm_mday, tm.tm_hour, tm.tm_min, tm.tm_sec,
+    tm.tm_gmtoff < 0 ? '-':'+', (int)abs(tm.tm_gmtoff)/3600, (int)(abs(tm.tm_gmtoff)/60)%60);
   rc = decode_utc(&pa,"TTL",NULL,str,0);
   fail_unless(rc == 0);
   memset(&pa,0,sizeof(pa));
   tm.tm_gmtoff -= 1800;
-  sprintf(str,"%04d-%02d-%02dT%02d:%02d:%02d%c%02d%02d",tm.tm_year+1900,tm.tm_mon+1,tm.tm_mday,tm.tm_hour,tm.tm_min,tm.tm_sec,tm.tm_gmtoff < 0 ? '-':'+',abs(tm.tm_gmtoff)/3600,(abs(tm.tm_gmtoff)/60)%60);
+  sprintf(str,"%04d-%02d-%02dT%02d:%02d:%02d%c%02d%02d",
+    tm.tm_year+1900, tm.tm_mon+1, tm.tm_mday, tm.tm_hour, tm.tm_min, tm.tm_sec,
+    tm.tm_gmtoff < 0 ? '-':'+', (int)abs(tm.tm_gmtoff)/3600, (int)(abs(tm.tm_gmtoff)/60)%60);
   rc = decode_utc(&pa,"TTL",NULL,str,0);
   fail_unless(rc == 0);
   now -= 90000;
@@ -40,11 +46,15 @@ START_TEST(test_one)
   rc = decode_utc(&pa,"TTL",NULL,str,0);
   fail_unless(rc == PBSE_BAD_UTC_RANGE);
   memset(&pa,0,sizeof(pa));
-  sprintf(str,"%04d-%02d-%02dT%02d:%02d:%02d%c%02d",tm.tm_year+1900,tm.tm_mon+1,tm.tm_mday,tm.tm_hour,tm.tm_min,tm.tm_sec,tm.tm_gmtoff < 0 ? '-':'+',abs(tm.tm_gmtoff)/3600);
+  sprintf(str,"%04d-%02d-%02dT%02d:%02d:%02d%c%02d",
+    tm.tm_year+1900, tm.tm_mon+1, tm.tm_mday, tm.tm_hour, tm.tm_min, tm.tm_sec,
+    tm.tm_gmtoff < 0 ? '-':'+', (int)abs(tm.tm_gmtoff)/3600);
   rc = decode_utc(&pa,"TTL",NULL,str,0);
   fail_unless(rc == PBSE_BAD_UTC_RANGE);
   memset(&pa,0,sizeof(pa));
-  sprintf(str,"%04d-%02d-%02dT%02d:%02d:%02d%c%02d%02d",tm.tm_year+1900,tm.tm_mon+1,tm.tm_mday,tm.tm_hour,tm.tm_min,tm.tm_sec,tm.tm_gmtoff < 0 ? '-':'+',abs(tm.tm_gmtoff)/3600,(abs(tm.tm_gmtoff)/60)%60);
+  sprintf(str,"%04d-%02d-%02dT%02d:%02d:%02d%c%02d%02d",
+    tm.tm_year+1900, tm.tm_mon+1, tm.tm_mday, tm.tm_hour, tm.tm_min, tm.tm_sec,
+    tm.tm_gmtoff < 0 ? '-':'+', (int)abs(tm.tm_gmtoff)/3600, (int)(abs(tm.tm_gmtoff)/60)%60);
   rc = decode_utc(&pa,"TTL",NULL,str,0);
   fail_unless(rc == PBSE_BAD_UTC_RANGE);
   }
